Move graphics portion of project into graphics folder
