+++ YoutubeOpenGL.vcxproj.filters
@@ -1,42 +1,29 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;c++;cppm;ixx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;h++;hm;inl;inc;ipp;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Source Files\imgui">
-      <UniqueIdentifier>{7a80e7e7-0fe1-48c1-800e-7ae85c69d5c8}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\imgui">
-      <UniqueIdentifier>{48e5ad9e-2275-4ef5-997d-4a1ac19e7062}</UniqueIdentifier>
-    </Filter>
+    <ClCompile Include="glad.c" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="imgui\*.cpp" />
+    <ClCompile Include="src\*.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="glad.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="imgui\*.cpp">
-      <Filter>Source Files\imgui</Filter>
-    </ClCompile>
-    <ClCompile Include="src\*.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="imgui\*.h" />
+    <ClInclude Include="src\*.h" />
+    <ClInclude Include="src\*.h" />
+    <ClInclude Include="src\*.h" />
+    <ClInclude Include="src\*.h" />
   </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="imgui\*.h">
-      <Filter>Header Files\imgui</Filter>
-    </ClInclude>
-    
-    <ClInclude Include="src\*.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
 </Project>
