+++ src/graphics/Camera.h
@@ -9,7 +9,7 @@
 #include<glm/gtx/rotate_vector.hpp>
 #include<glm/gtx/vector_angle.hpp>
 
-#include"shaderClass.h"
+#include"graphics/shaderClass.h"
 
 class Camera
 {
