+++ src/graphics/GUI.h
@@ -5,7 +5,7 @@
 #include"imgui.h"
 #include"imgui_impl_glfw.h"
 #include"imgui_impl_opengl3.h"
-#include"Scene.h"
+#include"graphics/Scene.h"
 class GUI{
     protected:
         Scene* scene;
