+++ src/graphics/Mesh.cpp
@@ -1,4 +1,4 @@
-#include "Mesh.h"
+#include "graphics/Mesh.h"
 #include<glm/glm.hpp>
 #include"util.h"
 #include <iostream>
