+++ src/graphics/Texture.h
@@ -4,7 +4,7 @@
 #include<glad/glad.h>
 #include<stb/stb_image.h>
 
-#include"shaderClass.h"
+#include"graphics/shaderClass.h"
 
 class Texture
 {
