+++ src/graphics/shaderClass.cpp
@@ -1,4 +1,4 @@
-#include"shaderClass.h"
+#include"graphics/shaderClass.h"
 #include<stdio.h>
 // Reads a text file and outputs a string with everything in the text file
 std::string get_file_contents(const char* filename)
